Re-structure inline with Docker's Official Images guidelines
In order for this repository to reach acceptance standards for
multi-arch in Docker's Official Images repo the guidelines [0]
need to be followed.

This change adapts the directory-level sorting order from
<version>-<architecture>-<base> to <version>-<base>-<architecture>
where <architecture> can be omitted from the latter structure when
a Dockerfile supports multiple architectures.

A good example of that is the Dockerfile for 3.0 based on Alpine,
which can support both ARM64v8 (AArch64) and x86_64 (and probably
some other architectures too [untested]) without changes.

[0] https://github.com/docker-library/official-images#multiple-architectures

Signed-off-by: Lee Jones <lee.jones@linaro.org>